+++ src/main/java/com/training/web/pages/flipkart/ContactUsPage.java
@@ -1,22 +1,27 @@
 package com.training.web.pages.flipkart;
 
 import com.training.basepages.FlipkartBasePage;
+
 import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selenide.$$x;
 import static com.codeborne.selenide.Selenide.$x;
+
 public class ContactUsPage extends FlipkartBasePage {
 
+    private String postalAddressLink = "//span[text()='Postal Address']";
+    private String addressBox = "//div[@class='_14_ipq']";
+
     public ContactUsPage clickPostalAddress() {
-        $x("//span[text()='Postal Address']").shouldBe(visible).click();
+        $x(postalAddressLink).shouldBe(visible).click();
         return this;
     }
 
     public String getPostalAddress() {
-       return  $$x("//div[@class='_14_ipq']").get(1).getText().replaceAll("[, \n]","");
+        return $$x(addressBox).get(1).getText().replaceAll("[, \n]", "");
     }
 
     public String getCorporateAddress() {
-        return  $$x("//div[@class='_14_ipq']").get(0).getText().replaceAll("[, \n]","");
+        return $$x(addressBox).get(0).getText().replaceAll("[, \n]", "");
 
     }
 }
